# docs/openapi.yml
openapi: 3.0.3
info:
  title: LotR Wiki API
  description: "API contract version v1.

    "
  version: 0.5.0
servers:
  - url: http://localhost:8000/api/v1
    description: Local dev server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained during login or registration.
...
  schemas:
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: Error code (for example, VALIDATION_ERROR)
            message:
              type: string
              description: Human-readable description
            fields:
              type: object
              nullable: true
              additionalProperties:
                type: array
                items:
                  type: string
              description: Field-level errors, if applicable
...
    PageSummary:
      type: object
      required:
        - slug
        - type
        - name
      properties:
        slug:
          type: string
        type:
          type: string
          description: Page type (character, race, location, ...)
        name:
          type: string
        image_url:
          type: string
          nullable: true
...
    NeighborsResponse:
      type: object
      properties:
        root_node:
          $ref: "#/components/schemas/PageSummary"
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/PageSummary"
        edges:
          type: array
          items:
            type: object
            properties:
              from: { type: string }
              to: { type: string }
              type: { type: string }
              properties:
                type: object
                additionalProperties: true
        stats:
          type: object
          properties:
            total_neighbors: { type: integer }
            by_type:
              type: object
              additionalProperties:
                type: integer
            by_relation:
              type: object
              additionalProperties:
                type: integer
...
paths:
...
  /analytics/neighbors:
    get:
      tags: [analytics]
      summary: Node neighbors graph
      operationId: neighbors
      parameters:
        - name: slug
          in: query
          required: true
          schema:
            type: string
        - name: through_nodes
          in: query
          required: false
          schema:
            type: string
          description: Comma-separated node types to include (all if empty)
        - name: through_rels
          in: query
          required: false
          schema:
            type: string
          description: Comma-separated relationship types (all if empty)
        - name: depth
          in: query
          schema:
            type: integer
            enum: [1, 2]
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            default: 100
            maximum: 1000
      responses:
        "200":
          description: Neighbors graph
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NeighborsResponse"
        "400":
          description: slug not provided
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Node not found
...